json problem fixed

+++ src/main/java/de/telekom/sea7/inter/repository/BicRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import de.telekom.sea7.impl.repository.Bic;
 
-public interface BicRepository extends JpaRepository<Bic, Integer> {
+public interface BicRepository extends JpaRepository<Bic, Long> {
 
 }
