+++ src/main/java/de/telekom/sea7/inter/repository/IbanRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import de.telekom.sea7.impl.repository.Iban;
 
-public interface IbanRepository extends JpaRepository<Iban, Integer> {
+public interface IbanRepository extends JpaRepository<Iban, Long> {
 
 }
